modified half error >> convert float

+++ xdecoder/architectures/xdecoder_model.py
@@ -522,19 +522,19 @@ class GeneralizedXdecoder(nn.Module):
 
     def semantic_inference(self, mask_cls, mask_pred, keep_sem_bgd=False):
         if keep_sem_bgd:
-            mask_cls = F.softmax(mask_cls, dim=-1)
+            mask_cls = F.softmax(mask_cls.float(), dim=-1)
         else:
-            mask_cls = F.softmax(mask_cls, dim=-1, dtype=torch.float)[..., :-1]
+            mask_cls = F.softmax(mask_cls.float(), dim=-1, dtype=torch.float)[..., :-1]
         # changed
         mask_pred = mask_pred.type(torch.float)
         # print(type(mask_pred))
-        mask_pred = mask_pred.sigmoid()
+        mask_pred = mask_pred.float().sigmoid()
         semseg = torch.einsum("qc,qhw->chw", mask_cls, mask_pred)
         return semseg
 
     def panoptic_inference(self, mask_cls, mask_pred):
-        scores, labels = F.softmax(mask_cls, dim=-1).max(-1)
-        mask_pred = mask_pred.sigmoid()
+        scores, labels = F.softmax(mask_cls.float(), dim=-1).max(-1)
+        mask_pred = mask_pred.float().sigmoid()
 
         keep = labels.ne(self.sem_seg_head.num_classes) & (scores > self.object_mask_threshold)
         cur_scores = scores[keep]
@@ -594,7 +594,7 @@ class GeneralizedXdecoder(nn.Module):
         image_size = mask_pred.shape[-2:]
 
         # [Q, K]
-        scores = F.softmax(mask_cls, dim=-1)[:, :-1]
+        scores = F.softmax(mask_cls.float(), dim=-1)[:, :-1]
         labels = torch.arange(self.sem_seg_head.num_classes, device=self.device).unsqueeze(0).repeat(self.num_queries, 1).flatten(0, 1)
         # scores_per_image, topk_indices = scores.flatten(0, 1).topk(self.num_queries, sorted=False)
         scores_per_image, topk_indices = scores.flatten(0, 1).topk(self.test_topk_per_image, sorted=False)
@@ -632,7 +632,7 @@ class GeneralizedXdecoder(nn.Module):
             result.pred_boxes = Boxes(torch.zeros(mask_pred.size(0), 4))
 
         # calculate average mask prob
-        mask_scores_per_image = (mask_pred.sigmoid().flatten(1) * result.pred_masks.flatten(1)).sum(1) / (result.pred_masks.flatten(1).sum(1) + 1e-6)
+        mask_scores_per_image = (mask_pred.float().sigmoid().flatten(1) * result.pred_masks.flatten(1)).sum(1) / (result.pred_masks.flatten(1).sum(1) + 1e-6)
         result.scores = scores_per_image * mask_scores_per_image
         result.pred_classes = labels_per_image
